refactor(D-168 A3): extract conversation_store from routes.py (verbatim)
- conversation_store.py: shared httpx client lifecycle (conversation-service) +
  _create_new_conversation + _lazy_import_history_with_retry + _ensure_conversation
  + _persist_assistant_message (D-006 ownership semantics untouched)
- routes.py re-imports the 6 moved names; main.py's close_conv_service_client
  import keeps resolving through routes
- manifest: API_SOURCE_FILES += conversation_store.py

routes.py: 2,465 -> 2,110 lines. Gates green; 486 affected tests pass.
