+++ microservices/orchestrator_service/src/api/api_sources.py
@@ -26,6 +26,7 @@ API_SOURCE_FILES: tuple[str, ...] = (
     "microservices/orchestrator_service/src/api/chat_context.py",  # D-168 Slice A1
     "microservices/orchestrator_service/src/api/identity_access.py",  # D-168 Slice A2
     "microservices/orchestrator_service/src/api/stream_serialization.py",  # D-168 Slice A2
+    "microservices/orchestrator_service/src/api/conversation_store.py",  # D-168 Slice A3
 )
